+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 - Add `-NoReset` switch parameter to skip service restart in PowerShell
   registration functions.
+- Support for [`NLog`](https://www.nuget.org/packages/NLog/)
+  logs instrumentation for versions `5.*` and `6.*` on .NET using duck typing
+  for zero-config auto-injection.
 
 ### Changed